@@ -3,7 +3,7 @@
   "version": "0.11.1",
   "description": "Install missing TypeScript typings for your dependencies.",
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=16.0.0"
   },
   "files": [
     "lib",
@@ -16,7 +16,7 @@
   "scripts": {
     "clean": "rimraf lib",
     "test": "jest",
-    "lint": "tslint --project tsconfig.json --fix 'src/**/*.ts' && tsc --noEmit && yarn format",
+    "lint": "eslint --fix 'src/**/*.ts' && tsc --noEmit && npm run format",
     "format": "prettier --write 'src/**/*.ts'",
     "lint:watch": "nodemon --exec npm run lint -e ts,js",
     "cover": "jest --coverage",
@@ -31,7 +31,7 @@
   },
   "lint-staged": {
     "*.ts": [
-      "tslint --project tsconfig.json --fix",
+      "eslint --fix",
       "prettier --write",
       "git add"
     ]
@@ -61,14 +61,15 @@
     "@types/jest": "^29.5.11",
     "@types/npm-registry-fetch": "^8.0.7",
     "@types/semver": "^7.5.6",
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "eslint": "^8.51.0",
+    "eslint-config-prettier": "^9.0.0",
     "jest": "^29.7.0",
     "nodemon": "^3.0.2",
     "prettier": "^3.1.1",
     "rimraf": "^5.0.5",
     "ts-jest": "^29.1.1",
-    "tslint": "^6.1.3",
-    "tslint-config-prettier": "^1.18.0",
-    "tslint-config-standard": "^9.0.0",
     "typescript": "^5.3.3"
   },
   "jest": {
